更新数据

+++ Chapter 3/mybatis-demo/src/main/java/geektime/spring/data/mybatisdemo/mapper/CoffeeMapper.java
@@ -14,7 +14,7 @@ public interface CoffeeMapper {
     @Insert("insert into t_coffee (name, price, create_time, update_time)"
             + "values (#{name}, #{price}, now(), now())")
     @Options(useGeneratedKeys = true)
-    Long save(Coffee coffee);
+    int save(Coffee coffee);
 
     @Select("select * from t_coffee where id = #{id}")
     @Results({
